pyproject.toml
[project]
name = "mini-genie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.1",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "wandb>=0.23.1",
]
